update storybook file for NFT card
